+++ src/ScanJob.py
@@ -38,7 +38,7 @@ from banglogging import log
 import banglogging
 from FileResult import FileResult
 from FileContentsComputer import *
-from Unpacker import *
+from UnpackManager import *
 from UnpackParserException import UnpackParserException
